safari-600.7-branch/LayoutTests/storage/websql/multiple-transactions.html

<html>
<head>
<script src="resources/database-common.js"></script>
<script src="multiple-transactions.js"></script>
<head>
<body onload="setupAndRunTest();">
This is a test to see if the database API allows multiple transactions to be queued on the same database at once:<br>
<pre id="console"> 
FAILURE: test didn't run.
</pre> 
</body>
</html>
