Safari-609.1.11.1/LayoutTests/storage/websql/multiple-transactions-expected.txt

CONSOLE MESSAGE: line 8: Web SQL is deprecated. Please use IndexedDB instead.
This is a test to see if the database API allows multiple transactions to be queued on the same database at once:
Transaction 1 Started
Transaction 1 Succeeded
Transaction 2 Started
Transaction 2 Succeeded
